Remove the rumor heat mechanism.
The rumor heat mechanism described in the Xerox paper that our rumor
implementation is based on uses the heat to determine how many times
a rumor has been sent to the network as a whole. Our implementation
didn't do that. Instead, it used the heat to determine how many times
a rumor had been sent to each member of the network.

It turns out that's fairly wasteful because we're using zmq as
a transport for the rumors, which uses TCP under the hood, so delivering
a rumor once was sufficient.

Now that we're switching to a mechanism that refreshes rumors
periodically, the heat mechanism doesn't serve a purpose. Instead, we
rely on a mechanism that's closer to active anti-entropy to ensure that
all members receive the rumors.

This commit also adds a refresh mechanism to rumors.

Signed-off-by: Josh Black <raskchanky@gmail.com>